requirements.txt

scapy>=2.5.0
Flask>=2.3.0
Flask-SocketIO>=5.3.0
Flask-CORS>=4.0.0
streamlit>=1.28.0
plotly>=5.15.0
pandas>=2.0.0
requests>=2.31.0
netifaces>=0.11.0

PyYAML>=6.0
